+++ apps/app/src/hooks/usePermissions.ts
@@ -1,0 +1,37 @@
+import { usePhoton } from '@photonhealth/react';
+import jwtDecode from 'jwt-decode';
+import { Permission } from 'packages/sdk/dist/types';
+import { useEffect, useState } from 'react';
+
+function checkHasPermission(subset: Permission[], superset: Permission[]) {
+  for (let i = 0; i < subset.length; i++) {
+    if (superset.indexOf(subset[i]) === -1) {
+      return false;
+    }
+  }
+  return true;
+}
+
+const usePermissions = (permissions: Permission[]) => {
+  const [userPermissions, setUserPermissions] = useState<Permission[]>([]);
+  const { getToken } = usePhoton();
+
+  useEffect(() => {
+    const getPermissions = async () => {
+      try {
+        const token = await getToken();
+        const decoded: { permissions: Permission[] } = jwtDecode(token);
+        setUserPermissions(decoded?.permissions || []);
+      } catch {
+        console.error('Something went wrong decoding user permissions');
+        setUserPermissions([]);
+      }
+    };
+
+    getPermissions();
+  }, []);
+
+  return checkHasPermission(permissions, userPermissions);
+};
+
+export default usePermissions;
