--- packages/elements/types.ts
@@ -1,15 +1,0 @@
-export type Permission =
-  | 'read:patient'
-  | 'write:patient'
-  | 'read:prescription'
-  | 'write:prescription'
-  | 'read:order'
-  | 'write:order'
-  | 'read:allergy'
-  | 'read:client'
-  | 'read:webhook_config'
-  | 'write:webhook_config'
-  | 'update:client_keys'
-  | 'read:organization'
-  | 'create:invite'
-  | 'write:roles';
